Cleanup many docstrings in ga.py
Unfortunately napoleon doesn't allow us to group our members together as was intended here - so we have to abandon numpydoc in this class and use `.. attribute` directly, which is quite verbose. Also frustratingly, we cannot use headings inside `Ga.__doc__`, so have to make do with `..rubric::`.

This also adds backticks and latex rendering in many places where it aids readability.