Make `Mv._make_x` functions private and static
These don't need to mutate any state, and they way they did before made the iterative approach in make_multivector quite confusing.

The functions now return a raw sympy expression, which is assigned to `self.obj` in a single place.

This also adds errors to the paths that didn't set anything at all.

Since these are now static methods, it is easier to remove the dictionary so that they can be called the usual way.

The one assignment to `self.grade_i` was pointless, as it was overwritten later anyway.