Remove non-existent magic methods
`__add_ab__` appeared in a very old `python-dev` discussion, but as far as I can tell has never been respected, and certainly is not respected in any supported version of python - the correct spelling is `__iadd__`.

Given that `galgebra` is working fine without these methods, and sympy does not define any in place operators, it seems wise to just remove them.