Remove assignments to mv.obj
It is much easier to reason about multivector objects if they are are not mutated unexpectedly.

This changes the following functions to not mutate their arguments. Any code relying solely on the return value will be unaffected:

* `Metric.applymv`
* `Mv.diff`
* `Mv.simplify`
* `Mv.expand`
* `Mv.collect`

This also changes `Mv.Mv_str` and `Mv.Mv_latex_str` to not mutate their multivectors when they are printed - this is super surprising behavior