Fix heterogeneous sdop.Add
Before this change, any addition other than `Sdop + Sdop` would raise a NameError, since the variable `coef` is not defined in the `elif` and `else` branch.

The intent of the old code was fairly clear - it was to insert a multiplicative term for `sdop2.Ga.Pdop_identity`, such that `(sdop + 2)(x) == sdop(x) + 2*x`, which is reflected by the repr in a shell.
This change just copies the code from `Dop`, which succeeds in doing exactly the same thing

This also adds a bunch more tests for both Sdop and Dop operations