Fix Pdop and Sdop shorthands
The previous code was nonsense - `len(args[0]) == 1 and isinstance(args[0],Symbol)` was never true, as `Symbol.__len__` is undefined.

This also makes it illegal to pass extra ignored arguments to `Pdop`.

Finally, this deprecates ``Pdop({}, ga=ga)` as an alias for `Pdop({}, ga=ga)` - there's no reason to have two ways to spell the same thing, and the second is more natural.