Deprecate the Pdop.Ga and Sdop.Ga attributes
These are scalar operators, they should know nothing about geomtric algebras.

For compatibility, we leave the `Ga` attribute behind, but do nothing other than propagate it in case the user is relying on it.
Uses of the attribute in user code will now warn.

In a later release, we can remove the attribute entirely.