Remove Dop.blade_rep
This function doesn't do anything, and causes LGTM warnings.
API surface that doesn't work is just confusing to users.

The intention of this function was probably something like:
```python
def blade_rep(self):
    return Dop(_consolidate_terms(
        (coef.blade_rep(), pdiff) for coef, pdiff in self.terms
    ), cmpflg=self.cmpflg, ga=self.Ga)
```

However, constructing a Dop with non-blade coefficients is unlikely in the first place, so if we add this function it is very unlikely that anyone would need it.

Closes gh-147