Remove trailing empty return statements
A return statement on the last line of a function, with no value to return, is a no-op.

There's already a very clear indicator that the end of the function has been reached anyway - the indentation level.

Found by looking for `^        return$`, and manually vetting each one for readability.
A few in nested functions remain.