Cleanup the very weird `Eprint` class which does not behave like a class at all
All its methods and attributes were static, and the instances were never used.

This replaces the class `Eprint` with a single function `Eprint`, which just changes global settings on the `GaPrinter` object.

This also removes without deprecation the useless `debug` and `on` arguments, as well as all the members of Eprint.