Run flake8 against the source, with a very generous ignore list
In order to keep this diff small, this ignores every warning we currently have, rather than fixing them.
This is enough to prevent the style drifting further from flake8 unnoticed.

Future patches can aim to cut down the warnings at a file-by-file basis.

CI is set up so that if linting fails, no tests are run.