Remove the no-longer needed print reconfiguration in Ga.__init__
`Ga()` and `Sm()` end up calling `str(some_sympy_object)` in multiple places.
Previously, this went through our custom printer, which meant we had to temporarily disable latex printing.

Making `str(some_sympy_object)` go through our printer was a mistake that interfered with sympy internals, so a previous commit made it no longer do that.
As a result, we no longer need to touch the printer in `__init__`.

Even though we're already not relying on it, this code is dangerous:

* If `Ga(...)` fails, then the latex printer is left disabled
* It isn't threadsafe, constructing a `Ga` in one thread interferes with printing in another