Do not clobber sympy.Basic.__str__
Instead we modify `print`, because at least nothing within sympy should be depending on that.

This fixes `simplify(expr)` behaving differently depending on whether latex printing was enabled or disabled when `simplify(expr)` was called for the _first_ time.
This is a horrible bug, and you could argue we deserved it for doing horrible things to sympy internals.

Ideally we would not replace `print` either, but that would be an API change, so should come as a future commit.

The notebook changes are a result of our printer no longer affecting sympy's default sort order for terms.
It's not clear to me why the formatting of the matrix output changed, but it seems harmless in any case.