Partially revert "Fixed multivector grade decomposition function in ga.Ga and added"
This partially reverts commit fc598df56ac6573248a187d514262f6312b17f31.

This commit began a refactor of the internals of `Mlt`, but never finished, resulting in the existing documentation being wrong and the function being unusable.

We can explore re-introducing this change later, but for now we should restore it to its documented behavior.

Note the original commit contained a mixture of changes, and only the Mlt-specific ones needed reverting.