Improve the rendering of basis blades in the printer
This introduces new sympy objects to represent the basis blades and the dot product components of the metric, so that we can define printers for these specifically.

It also implements the sympy "pretty" printer for these types, which you can see in the plaintext of the printing-galgebra notebook.

The main notebook changes here are the addition of `\boldsymbol` around basis vectors in places where that printing was lost.

As a result of the new types of these symbols, the order of terms in printing changes in a few places (as this is based upon the class names). None of these are changes affect the meaning.