Fix the intended column alignment
The `obj is None` condition was clearly intended to be `obj is not None`, as `isinstance(None, dict)` would be nonsense.

With this change in mind, this code can be collapsed into a single call to `max`.

Fix the intended column alignment

The `obj is None` condition was clearly intended to be `obj is not None`, as `isinstance(None, dict)` would be nonsense.

With this change in mind, this code can be collapsed into a single call to `max`.