Move global settings out of GaLatexPrinter (#364)
The sympy way is to store these in a per-instance `_settings` dict.

This patch makes the assumption that the attributes of `GaLatexPrinter` were not public API, and that user code should only be accessing them through `Format`.

This means that the galgebra printer now respects the `inv_trig_style` argument to `sympy.init_printing`.