Eliminate the global `.fmt` attribute on printers
The juggling of saving and restoring this safe was not done properly.
In order to make `Fmt` work in an ipython notebook, we need it to return a special object that remembers its format settings.

Note that the examples that have changed are typically of the form:
```
print(x)
y.Fmt(...)  # does nothing!
print(z)
```
This would previously result in the print format of `z` being different to that of `x`, due to bad management of the global state.

This also makes `Fmt` work in plaintext mode.