Do not make any __str__ or __repr__ output latex
This follows on from #371, and also hitting `Sdop`, `Pdop`, `Lt`, and `Mlt`.
Note that this once again fixes the plaintext output to actually be plaintext.

This then moves all the duplicated code to `printer.py`.
As a result of this change, only one line of code outside of `printer.py` uses `GaLatexPrinter.latex_flg`, and it only does so as a workaround.

Note that if `__str__` is not defined, it defaults to `__repr__` - so we can save some lines by only defining the latter.