Change the latex printing of Mv to use aligned instead of align*
The former can be used in math mode, but the latter is only legal in text mode.
The sympy latex printer expects all results to be usable in math mode.

This allows us to remove lots of workarounds that attempt to solve this problem by moving `align` to the outside, which is annoying and error-prone.

After doing this, many of our functions become identical, so we move the methods to `printer.py` to reduce duplication.