Update notebook to the behavior from sympy@e31c4cdedb5080325b3fd04f4b4826d9dab65b26
The upstream change is in our favor, but it breaks our tests that were demonstrating previously-broken behavior.

For now, we tag those tests with `NBVAL_IGNORE_OUTPUT` to avoid them failing on released versions of sympy.

The other changes in this file are already covered by our nbval sanitization configuration.